+++ core/src/main/java/io/snabble/sdk/events/Events.kt
@@ -6,7 +6,6 @@ import android.os.SystemClock
 import io.snabble.sdk.Project
 import io.snabble.sdk.Shop
 import io.snabble.sdk.ShoppingCart
-import io.snabble.sdk.ShoppingCart.SimpleShoppingCartListener
 import io.snabble.sdk.Snabble
 import io.snabble.sdk.codes.ScannedCode
 import io.snabble.sdk.events.data.Event
@@ -20,6 +19,7 @@ import io.snabble.sdk.events.data.payload.PayloadProductNotFound
 import io.snabble.sdk.events.data.payload.PayloadSessionEnd
 import io.snabble.sdk.events.data.payload.PayloadSessionStart
 import io.snabble.sdk.shoppingcart.data.Taxation
+import io.snabble.sdk.shoppingcart.data.listener.SimpleShoppingCartListener
 import io.snabble.sdk.utils.DateUtils
 import io.snabble.sdk.utils.GsonHolder
 import io.snabble.sdk.utils.Logger
